+++ src/os/getCPU.js
