+++ src/os/getEOL.js
